Crate pallet_vanchor_verifier

VAnchor Verifier Module

A module for abstracting over many arbitrary zero-knowledge verifiers for arbitrary zero-knowledge gadgets splitting on vanchor length. This pallet should store verifying keys and any other verification specific parameters for different backends that we support in Webb’s ecosystem of runtime modules.

Overview

The VAnchor Verifier module provides functionality for zero-knowledge verifier management including:

Setting parameters for zero-knowledge verifiers

Setting the maintainer of the parameters

To use it in your runtime, you need to implement the verifier Config. Additionally, you will want to implement the verifier traits defined in the webb_primitives::verifier module.

The supported dispatchable functions are documented in the Call enum.

Terminology

Goals

The verifier system in Webb is designed to make the following possible:

Define.

Interface

Related Modules

System

Support

Re-exports

pub use weights::WeightInfo;

pub use pallet::*;

Modules

pallet

The module that hosts all the FRAME types needed to add this pallet to a runtime.

weights

Autogenerated weights for pallet_vanchor_verifier